requirements.txt
Django
pyjwt
celery
requests
backoff
pycrypto
git+https://github.com/TheEnbyperor/django-keycloak-auth